+++ src/language/general/runner.js
@@ -1,6 +1,6 @@
 const { cancellablePromise } = require("../../lib/cancellablePromise");
 const { getPositionOfFrom } = require("../../lib/getPositionOfFrom");
-const { promisable } = require("../../lib/pipePromise")
+const { promisable, promisableOne } = require("../../lib/pipePromise")
 
 module.exports.runner = function runner(languageRunner, editor, after) {
     let _languageRunner = null;
@@ -10,18 +10,20 @@ module.exports.runner = function runner(languageRunner, editor, after) {
 
         let hints = [];
         let nodes = null;
-        _languageRunner = (_nodes = null) => promisable(async () => await languageRunner((action) => {
+        let exclude = [];
+        _languageRunner = (_nodes = null, _exclude = null) => promisable(async () => await languageRunner((action) => {
             return promisable(action, () => state.done);
-        }, text, editor, positionOf, _nodes), () => state.done);
+        }, text, editor, positionOf, _nodes, _exclude), () => state.done);
 
 
         try {
-            [hints, nodes] = await _languageRunner().promise;
+            [hints, nodes, exclude] = await _languageRunner().promise;
             if (hints.length === 0) {
-                [hints, nodes] = await new Promise((r) => {
+                [hints, nodes, exclude] = await promisableOne((r) => {
                     let count = 0;
+                    console.log('retry', count);
                     let retry = () => setTimeout(async () => {
-                        let response = await _languageRunner(nodes).promise;
+                        let response = await _languageRunner(nodes, exclude).promise;
                         if (response.length && response[0].length || count > 3) {
                             r(response);
                         } else {
@@ -30,16 +32,16 @@ module.exports.runner = function runner(languageRunner, editor, after) {
                         }
                     }, 2000);
                     retry();
-                });
+                }, () => state.done).promise;
             }
 
             if (hints.length !== 0) {
                 after(hints);
             }
 
-            resolve([hints, nodes]);
+            resolve([hints, nodes, exclude]);
         } catch (e) {
-            resolve([hints, nodes]);
+            resolve([hints, nodes, exclude]);
         }
     });
     _runner.catch(e => { });