+++ src/language/javascript/providers/hover.js
@@ -1,24 +1,26 @@
 const vscode = require('vscode');
 const { TypescriptParser } = require('typescript-parser');
+const { hoverProvider } = require('../../general/providers');
 let typescriptParser = new TypescriptParser();
 module.exports.hoverProvider = async (editor, node, positionOf) => {
     let nodePosition = positionOf(node.start);
-    const hoverCommand = await vscode.commands.executeCommand(
-        "vscode.executeHoverProvider",
-        editor.document.uri,
-        nodePosition,
-    );
+
+    let hoverCommand = await hoverProvider(editor, node, nodePosition);
 
     if (hoverCommand.length > 0 && hoverCommand[0].contents && hoverCommand[0].contents.length > 0) {
         let res = hoverCommand[0].contents[0].value;
         if (res) {
+            let subparams;
+
             let parsingString = /```typescript(.*?)```/s.exec(res);
             if (!parsingString) {
                 return false;
             }
             let parsed = await typescriptParser.parseSource(parsingString[1]);
 
-            let subparams = parsed.declarations && parsed.declarations.length && parsed.declarations[0].parameters && parsed.declarations[0].parameters.map(e => e.name);
+            subparams = parsed.declarations && parsed.declarations.length && parsed.declarations[0].parameters && parsed.declarations[0].parameters.map(e => e.name);
+            parsed[res] = subparams;
+
             if (!subparams) {
                 return false;
             }
@@ -44,8 +46,10 @@ module.exports.hoverProvider = async (editor, node, positionOf) => {
                 }
                 params.push({
                     label: label.replace('?', '').trim() + ':',
-                    range: new vscode.Range(positionOf(node.arguments[i].start),
-                        positionOf(node.arguments[i].end))
+                    nodeStart: node.start,
+                    currentNodeStart: node.start,
+                    start: node.arguments[i].start,
+                    end: node.arguments[i].end
                 });
             }
             return params;